pyproject.toml
[project]
name = "housekeeper"
version = "1.0.0"
description = "Tool to crean up your laptop."
dependencies = [
    "argparse>=1.4.0",
    "datetime>=5.5",
    "pathlib>=1.0.1",
    "pyyaml>=6.0.1",
]
readme = "README.md"
requires-python = ">= 3.12"
license = { text = "MIT" }

[project.scripts]
"housekeep" = "housekeeper.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/housekeeper"]
